Lap_map_main/requirements.txt
numpy>=1.24.0
RPi.GPIO>=0.7.1
smbus2>=0.4.2
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
